Fix double restart when --watch is used
When `tns livesync <platform> --watch` is executed, the application is restarted twice as typescript's watch command is compiling the sources again.
CLI starts the watch command last, after the app is already started on the device. The watch command is compiling the .ts files to .js. CLI detects these changes and restarts the app.
In order to fix this, make the watch hook to return a promise. CLI will not start "watching" for file changes until the promise is resolved.
The watch hook's promise will be resolved when the following message is printed by TypeScript compiler: "Compilation complete. Watching for file changes.".