Do not overwrite typescript dependency
When you call `npm install --save-dev <package>` npm does not consider the current version of the package from devDependencies and always installs latest one.
So when the postinstall script of the `nativescript-dev-typescript` is executed, it always updates typescript dependency to the latest version.

Fix this by explicitly checking if there's already typescript dependency and do not call `npm install` in such case.