Remove usbLiveSyncService.isInitialized check
In case LiveSync process is currently running, `before-prepare` hook must not run the `tsc` compiler. We've been doing this by checking a special property from CLI - `usbLiveSyncService.isInitialized`. However, there's no need to use this service - we have the information if we have started `tsc` process, so just use it instead of relying on `usbLiveSyncService`.
This will allow us to remove the usbLiveSyncService from CLI as it is no longer used anywhere.